+++ tests/testthat/test-define_trait.R
@@ -6,7 +6,7 @@ make_tiny_pop <- function(pop_name = "t") {
     chr_len_Mb = 100,
     db_path    = ":memory:"
   ) |>
-    define_founder_haplotypes(n_haplotypes = 100, fixed_allele_freq = 0.5)
+    define_founder_haplotypes(n_haplotypes = 100, method = "fixed")
   pop <- add_founders(pop, n_males = 25, n_females = 25, line_name = "A")
   pop <- get_table(pop, "ind_meta") |> mutate_table(gen = 0L)
   pop